refactor: Move NoEvictScheduledBlocksManager and MaxUtilizationScheduledBlocksManager classes to separate header file
- Moved NoEvictScheduledBlocksManager and MaxUtilizationScheduledBlocksManager from kvCacheManager.h to a new scheduledBlocksManager.h.
- Updated capacityScheduler.cpp to include the new scheduledBlocksManager header.
