+++ cpp/include/tensorrt_llm/batch_manager/kvCacheManager.h
@@ -1668,90 +1668,4 @@ private:
     runtime::ITensor::SharedPtr mBlockScalePoolPointers;
 };
 
-class NoEvictScheduledBlocksManager
-{
-public:
-    explicit NoEvictScheduledBlocksManager(BaseKVCacheManager const& kvCacheManager)
-        : mKvCacheManager(kvCacheManager)
-        , mAvailableBlocks(mKvCacheManager.getBlockManager().getNumFreeBlocksPerWindowSize())
-    {
-    }
-
-    void decrementReservedBlocks(LlmRequest const& req)
-    {
-        for (auto& [windowSize, availableBlocks] : mAvailableBlocks)
-        {
-            availableBlocks -= mKvCacheManager.getRemainingBlocksToCompletion(req, windowSize);
-        }
-    }
-
-    bool enoughAvailableBlocks(LlmRequest const& req)
-    {
-        return std::all_of(mAvailableBlocks.cbegin(), mAvailableBlocks.cend(),
-            [this, &req](auto const& pair)
-            {
-                auto const& [windowSize, availableBlocks] = pair;
-                auto const neededBlocks = mKvCacheManager.getRemainingBlocksToCompletion(req, windowSize);
-                return neededBlocks <= availableBlocks;
-            });
-    }
-
-private:
-    BaseKVCacheManager const& mKvCacheManager;
-    std::map<SizeType32, SizeType32> mAvailableBlocks;
-};
-
-class MaxUtilizationScheduledBlocksManager
-{
-public:
-    MaxUtilizationScheduledBlocksManager(BaseKVCacheManager const& kvCacheManager, bool twoStepsLookAhead)
-        : mKvCacheManager(kvCacheManager)
-        , mTwoStepsLookAhead(twoStepsLookAhead)
-    {
-        auto const& windowSizes = mKvCacheManager.getBlockManager().getWindowSizesMetadata();
-        for (auto const& [windowSize, _] : windowSizes)
-        {
-            mNumScheduledBlocks[windowSize] = 0;
-        }
-    }
-
-    std::optional<std::map<SizeType32, SizeType32>> prepareNewNumberOfBlocksIfWeEndUpScheduling(LlmRequest const& req)
-    {
-        std::map<SizeType32, SizeType32> blocksIfScheduled;
-        for (auto const& [windowSize, numScheduled] : mNumScheduledBlocks)
-        {
-            auto const required = mKvCacheManager.getNeededBlocksOneStep(req, mTwoStepsLookAhead, windowSize);
-
-            TLLM_LOG_DEBUG("MaxUtilizationScheduler: request ID %lu required blocks %i for %i window size",
-                req.mRequestId, required, windowSize);
-
-            auto const scheduledTotal = numScheduled + required;
-            bool const hasFreeBlocks
-                = mKvCacheManager.getBlockManager().schedulingHasFreeBlocks(scheduledTotal, windowSize);
-            if (!hasFreeBlocks)
-            {
-                return std::nullopt;
-            }
-            blocksIfScheduled[windowSize] = scheduledTotal;
-        }
-        return blocksIfScheduled;
-    }
-
-    void updateScheduledBlocks(std::map<SizeType32, SizeType32> const& numBlocksIfScheduled)
-    {
-        assert(numBlocksIfScheduled.size() == mNumScheduledBlocks.size());
-        for (auto const& [windowSize, blocksIfScheduled] : numBlocksIfScheduled)
-        {
-            TLLM_LOG_DEBUG(
-                "MaxUtilizationScheduler: scheduled blocks %i for window size %i", blocksIfScheduled, windowSize);
-            mNumScheduledBlocks.at(windowSize) = blocksIfScheduled;
-        }
-    }
-
-private:
-    BaseKVCacheManager const& mKvCacheManager;
-    std::map<SizeType32, SizeType32> mNumScheduledBlocks;
-    bool const mTwoStepsLookAhead;
-};
-
 } // namespace tensorrt_llm::batch_manager::kv_cache_manager
